requirement.txt
pip==20.0.2   
beautifulsoup4==4.9.3    
dateformat==0.9.7    
Django==3.1.1    
django-cors-headers==3.5.0    
django-filter==2.3.0    
djangorestframework==3.11.1   
djangorestframework-simplejwt==4.4.0    
Markdown==3.2.2    
mysqlclient==2.0.1    
PyJWT==1.7.1    
pyOpenSSL==19.1.0   
pytz==2020.1   
requests==2.24.0   
service-identity==18.1.0   
setuptools==50.3.0   
six==1.15.0   
urllib3==1.25.10  
wheel==0.35.1   
